Move ClipboardButton out of components/common.js
