+++ src/components/UriViewer.js
@@ -3,8 +3,9 @@ import _ from 'lodash/fp'
 import { Fragment, useState } from 'react'
 import { div, h, img, input } from 'react-hyperscript-helpers'
 import { requesterPaysWrapper, withRequesterPaysHandler } from 'src/components/bucket-utils'
+import { ClipboardButton } from 'src/components/ClipboardButton'
 import Collapse from 'src/components/Collapse'
-import { ButtonPrimary, ClipboardButton, Link } from 'src/components/common'
+import { ButtonPrimary, Link } from 'src/components/common'
 import { FileProvenance } from 'src/components/data/data-table-provenance'
 import { getDownloadCommand, getUserProjectForWorkspace, parseGsUri } from 'src/components/data/data-utils'
 import { spinner } from 'src/components/icons'
