+++ src/components/workspace-utils.js
@@ -2,7 +2,8 @@ import debouncePromise from 'debounce-promise'
 import _ from 'lodash/fp'
 import { Fragment, useState } from 'react'
 import { b, div, h, p, span } from 'react-hyperscript-helpers'
-import { AsyncCreatableSelect, ButtonPrimary, ButtonSecondary, Clickable, ClipboardButton, IdContainer, Link, Select, spinnerOverlay } from 'src/components/common'
+import { ClipboardButton } from 'src/components/ClipboardButton'
+import { AsyncCreatableSelect, ButtonPrimary, ButtonSecondary, Clickable, IdContainer, Link, Select, spinnerOverlay } from 'src/components/common'
 import DelayedRender from 'src/components/DelayedRender'
 import { icon, spinner } from 'src/components/icons'
 import { ValidatedInput } from 'src/components/input'
