+++ src/pages/groups/List.js
@@ -1,7 +1,8 @@
 import _ from 'lodash/fp'
 import { Fragment, useEffect, useState } from 'react'
 import { a, div, h, h2 } from 'react-hyperscript-helpers'
-import { ButtonPrimary, ClipboardButton, DeleteConfirmationModal, HeaderRenderer, IdContainer, Link, spinnerOverlay } from 'src/components/common'
+import { ClipboardButton } from 'src/components/ClipboardButton'
+import { ButtonPrimary, DeleteConfirmationModal, HeaderRenderer, IdContainer, Link, spinnerOverlay } from 'src/components/common'
 import FooterWrapper from 'src/components/FooterWrapper'
 import { AdminNotifierCheckbox } from 'src/components/group-common'
 import { icon } from 'src/components/icons'
