+++ src/pages/workspaces/workspace/jobHistory/SubmissionDetails.js
@@ -4,7 +4,8 @@ import { Fragment, useEffect, useState } from 'react'
 import { div, h, h4 } from 'react-hyperscript-helpers'
 import { AutoSizer } from 'react-virtualized'
 import * as breadcrumbs from 'src/components/breadcrumbs'
-import { ButtonSecondary, ClipboardButton, Link, Select } from 'src/components/common'
+import { ClipboardButton } from 'src/components/ClipboardButton'
+import { ButtonSecondary, Link, Select } from 'src/components/common'
 import { centeredSpinner, icon } from 'src/components/icons'
 import { DelayedSearchInput } from 'src/components/input'
 import {
